Python Shell
First you need to open the script and change the host and port variable to your local or public ip and the port you want to listen on.
Then you can use pyinstaller to convert it to an exe file. It preferably that you use netcat for the listener and when the target executes the script you will have access to their computer.